Reorganize util package with functional grouping
