+++ src/main/java/com/nonxedy/nonchat/command/impl/IgnoreCommand.java
@@ -19,7 +19,7 @@ import org.jetbrains.annotations.NotNull;
 
 import com.nonxedy.nonchat.Nonchat;
 import com.nonxedy.nonchat.config.PluginMessages;
-import com.nonxedy.nonchat.util.ColorUtil;
+import com.nonxedy.nonchat.util.core.colors.ColorUtil;
 
 public class IgnoreCommand implements CommandExecutor, TabCompleter {
