+++ src/main/java/com/nonxedy/nonchat/core/BroadcastManager.java
@@ -12,8 +12,8 @@ import org.bukkit.scheduler.BukkitTask;
 
 import com.nonxedy.nonchat.Nonchat;
 import com.nonxedy.nonchat.config.PluginConfig;
-import com.nonxedy.nonchat.util.BroadcastMessage;
-import com.nonxedy.nonchat.util.ColorUtil;
+import com.nonxedy.nonchat.util.core.broadcast.BroadcastMessage;
+import com.nonxedy.nonchat.util.core.colors.ColorUtil;
 
 import net.kyori.adventure.text.Component;
