+++ src/main/java/com/nonxedy/nonchat/core/MessageManager.java
@@ -11,8 +11,8 @@ import com.nonxedy.nonchat.command.impl.IgnoreCommand;
 import com.nonxedy.nonchat.command.impl.SpyCommand;
 import com.nonxedy.nonchat.config.PluginConfig;
 import com.nonxedy.nonchat.config.PluginMessages;
-import com.nonxedy.nonchat.util.ChatTypeUtil;
-import com.nonxedy.nonchat.util.ColorUtil;
+import com.nonxedy.nonchat.util.chat.formatting.ChatTypeUtil;
+import com.nonxedy.nonchat.util.core.colors.ColorUtil;
 
 import net.kyori.adventure.text.Component;
