+++ src/main/java/com/nonxedy/nonchat/util/chat/filters/CapsFilter.java
@@ -1,4 +1,4 @@
-package com.nonxedy.nonchat.util;
+package com.nonxedy.nonchat.util.chat.filters;
 
 /**
  * Filters excessive capitalization in chat messages
