+++ src/main/java/com/nonxedy/nonchat/util/chat/formatting/MessageFormatter.java
@@ -1,6 +1,7 @@
-package com.nonxedy.nonchat.util;
+package com.nonxedy.nonchat.util.chat.formatting;
 
 import com.nonxedy.nonchat.config.PluginMessages;
+import com.nonxedy.nonchat.util.core.colors.ColorUtil;
 
 import net.kyori.adventure.text.Component;
