+++ src/main/java/com/nonxedy/nonchat/util/lang/LanguageManager.java
@@ -1,4 +1,4 @@
-package com.nonxedy.nonchat.util;
+package com.nonxedy.nonchat.util.lang;
 
 import java.io.File;
 import java.io.IOException;
@@ -9,6 +9,8 @@ import java.util.Map;
 
 import org.bukkit.configuration.file.FileConfiguration;
 import org.bukkit.configuration.file.YamlConfiguration;
+
+import com.nonxedy.nonchat.util.core.colors.ColorUtil;
 
 // Manages language files and provides methods to load and get translations
 public class LanguageManager {
